Re-compile stan models for each container run
Because of the linking to libtbb there is a mismatch between locally and
container compiled stan models. Instead of mucking around with rpath,
just be simple and re-compile everything when the container is run to
ensure a clean slate. Ideally, the pipeline is only executed in the
container once during a release anyways.